Add StringFileInfo #99
